FIXED Minor edit in AIUtil, added const to constant functions
FIXED Guards can attack the same Enemy without game crashing
FIXED Enemy constructor overloading
ADDED Enemies can't attack Guard if not InRange
REMOVED Unnecessary code in Guard's constructor

+++ StortSpelprojekt/StortSpelprojekt/ObjectHandler.h
@@ -55,7 +55,7 @@ private:
 	Grid* _buildingGrid;
 	Level::LevelHeader _currentLevelHeader;
 
-	int _idCount = 0;
+	unsigned short _idCount = 0;
 	int _objectCount = 0;
 	string _levelfolder;
